mod_tile_lite without rendering
This is a stripped-down version of mod_tile which ignores all the
complexity of rendering management and assumes that all tiles are
already in cache.

I also removed the (IMHO wrong) parsing of the .ini file as it doesn't
really make sense that part of the *Apache HTTPD* configuration is
stored in the renderd configuration. File serving/map configuration and
http bindings are different concepts.